requirements.txt
amqp==5.0.6
appnope==0.1.2
argh==0.26.2
asgiref==3.3.4
backcall==0.2.0
billiard==3.6.4.0
celery==5.0.5
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.1.6
decorator==5.0.7
Django==3.2
django-admin-sortable2==1.0
django-emoji-picker==0.0.6
django-extensions==3.1.3
django-shell-plus==1.1.7
djangorestframework==3.12.4
gtp==0.2
gunicorn==20.1.0
idna==2.10
ipython==7.22.0
ipython-genutils==0.2.0
jedi==0.18.0
kombu==5.0.2
parso==0.8.2
pexpect==4.8.0
pickleshare==0.7.5
prompt-toolkit==3.0.18
psycopg2==2.8.6
ptyprocess==0.7.0
Pygments==2.8.1
pyTelegramBotAPI==3.7.7
pytz==2021.1
PyYAML==5.4.1
requests==2.25.1
sgf==0.5
six==1.15.0
sqlparse==0.4.1
traitlets==5.0.5
urllib3==1.26.4
vine==5.0.0
watchdog==2.0.2
wcwidth==0.2.5
